chore: move node test utils from memory package to utils package
This moves the common code to bring up nodes in tests to the utils
package and updates all references to it. This is a refactor to prepare
for the final removal of the memory package.
