+++ deployment/environment/memory/aptos_chains.go
@@ -12,7 +12,6 @@ import (
 	cldf_aptos "github.com/smartcontractkit/chainlink-deployments-framework/chain/aptos"
 	cldf_aptos_provider "github.com/smartcontractkit/chainlink-deployments-framework/chain/aptos/provider"
 	"github.com/smartcontractkit/chainlink/v2/core/logger"
-	"github.com/smartcontractkit/chainlink/v2/core/services/chainlink"
 )
 
 func getTestAptosChainSelectors() []uint64 {
@@ -48,23 +47,6 @@ func generateChainsAptos(t *testing.T, numChains int) []cldf_chain.BlockChain {
 	return chains
 }
 
-func createAptosChainConfig(chainID string, chain cldf_aptos.Chain) chainlink.RawConfig {
-	chainConfig := chainlink.RawConfig{}
-
-	chainConfig["Enabled"] = true
-	chainConfig["ChainID"] = chainID
-	chainConfig["NetworkName"] = "localnet"
-	chainConfig["NetworkNameFull"] = "aptos-localnet"
-	chainConfig["Nodes"] = []any{
-		map[string]any{
-			"Name": "primary",
-			"URL":  chain.URL,
-		},
-	}
-
-	return chainConfig
-}
-
 func migrateAccountToFA(t *testing.T, signer aptos.TransactionSigner, client aptos.AptosRpcClient) error {
 	// Migrate APT Coin to FA, required for CCIP
 	payload := aptos.TransactionPayload{
@@ -96,18 +78,6 @@ func migrateAccountToFA(t *testing.T, signer aptos.TransactionSigner, client apt
 	accountAddress := signer.AccountAddress()
 	logger.TestLogger(t).Infof("Migrated account %v to Fungible Asset APT", accountAddress.StringLong())
 	return err
-}
-
-func fundNodesAptos(t *testing.T, aptosChain cldf_aptos.Chain, nodes []*Node) {
-	for _, node := range nodes {
-		aptoskeys, err := node.App.GetKeyStore().Aptos().GetAll()
-		require.NoError(t, err)
-		require.Len(t, aptoskeys, 1)
-		transmitter := aptoskeys[0]
-		transmitterAccountAddress := aptos.AccountAddress{}
-		require.NoError(t, transmitterAccountAddress.ParseStringRelaxed(transmitter.Account()))
-		FundAptosAccount(t, aptosChain.DeployerSigner, transmitterAccountAddress, 100*1e8, aptosChain.Client)
-	}
 }
 
 func FundAptosAccount(t *testing.T, signer aptos.TransactionSigner, to aptos.AccountAddress, amount uint64, client aptos.AptosRpcClient) {
