+++ deployment/environment/memory/environment.go
@@ -8,11 +8,11 @@ import (
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/core/types"
 
-	"github.com/smartcontractkit/freeport"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap/zapcore"
 
 	"github.com/smartcontractkit/chainlink/deployment"
+	"github.com/smartcontractkit/chainlink/deployment/utils/nodetestutils"
 
 	"github.com/smartcontractkit/chainlink-common/pkg/logger"
 
@@ -58,17 +58,6 @@ const (
 var ContractVersionShortSha = map[CCIPSolanaContractVersion]string{
 	SolanaContractV0_1_0: "0ee732e80586",
 	SolanaContractV0_1_1: "7f8a0f403c3a",
-}
-
-type NewNodesConfig struct {
-	LogLevel zapcore.Level
-	// BlockChains to be configured
-	BlockChains    cldf_chain.BlockChains
-	NumNodes       int
-	NumBootstraps  int
-	RegistryConfig deployment.CapabilityRegistryConfig
-	// SQL queries to run after DB creation, typically used for setting up testing state. Optional.
-	CustomDBSetup []string
 }
 
 // For placeholders like aptos
@@ -121,72 +110,11 @@ func NewMemoryChainsTron(t *testing.T, numChains int) []cldf_chain.BlockChain {
 	return generateChainsTron(t, numChains)
 }
 
-func NewNodes(
-	t *testing.T,
-	cfg NewNodesConfig,
-	configOpts ...ConfigOpt,
-) map[string]Node {
-	nodesByPeerID := make(map[string]Node)
-	if cfg.NumNodes+cfg.NumBootstraps == 0 {
-		return nodesByPeerID
-	}
-	ports := freeport.GetN(t, cfg.NumNodes+cfg.NumBootstraps)
-	// bootstrap nodes must be separate nodes from plugin nodes,
-	// since we won't run a bootstrapper and a plugin oracle on the same
-	// chainlink node in production.
-	for i := 0; i < cfg.NumBootstraps; i++ {
-		// TODO: bootstrap nodes don't have to support anything other than the home chain.
-		// We should remove all non-home chains from the config below and make sure things
-		// run smoothly.
-		c := NewNodeConfig{
-			Port:           ports[i],
-			BlockChains:    cfg.BlockChains,
-			LogLevel:       cfg.LogLevel,
-			Bootstrap:      true,
-			RegistryConfig: cfg.RegistryConfig,
-			CustomDBSetup:  cfg.CustomDBSetup,
-		}
-		node := NewNode(t, c, configOpts...)
-		nodesByPeerID[node.Keys.PeerID.String()] = *node
-		// Note in real env, this ID is allocated by JD.
-	}
-	var nodes []*Node
-	for i := range cfg.NumNodes {
-		c := NewNodeConfig{
-			Port:           ports[cfg.NumBootstraps+i],
-			BlockChains:    cfg.BlockChains,
-			LogLevel:       cfg.LogLevel,
-			Bootstrap:      false,
-			RegistryConfig: cfg.RegistryConfig,
-			CustomDBSetup:  cfg.CustomDBSetup,
-		}
-		// grab port offset by numBootstraps, since above loop also takes some ports.
-		node := NewNode(t, c, configOpts...)
-		nodesByPeerID[node.Keys.PeerID.String()] = *node
-		// Note in real env, this ID is allocated by JD.
-
-		nodes = append(nodes, node)
-	}
-
-	// Funding (only non-bootstrap nodes)
-	for _, tonChain := range cfg.BlockChains.TonChains() {
-		fundNodesTon(t, tonChain, nodes)
-	}
-	for _, aptosChain := range cfg.BlockChains.AptosChains() {
-		fundNodesAptos(t, aptosChain, nodes)
-	}
-	for _, solChain := range cfg.BlockChains.SolanaChains() {
-		fundNodesSol(t, solChain, nodes)
-	}
-
-	return nodesByPeerID
-}
-
 func NewMemoryEnvironmentFromChainsNodes(
 	ctx func() context.Context,
 	lggr logger.Logger,
 	blockchains cldf_chain.BlockChains,
-	nodes map[string]Node,
+	nodes map[string]nodetestutils.Node,
 ) cldf.Environment {
 	var nodeIDs []string
 	for id := range nodes {
@@ -233,7 +161,7 @@ func NewMemoryEnvironment(
 		slices.Concat(evmChains, solChains, aptosChains, zkChains, suiChains, tonChains, tronChains),
 	)
 
-	c := NewNodesConfig{
+	c := nodetestutils.NewNodesConfig{
 		LogLevel:       logLevel,
 		BlockChains:    chains,
 		NumNodes:       config.Nodes,
@@ -241,7 +169,7 @@ func NewMemoryEnvironment(
 		RegistryConfig: config.RegistryConfig,
 		CustomDBSetup:  config.CustomDBSetup,
 	}
-	nodes := NewNodes(t, c)
+	nodes := nodetestutils.NewNodes(t, c)
 	var nodeIDs []string
 	for id, node := range nodes {
 		require.NoError(t, node.App.Start(t.Context()))
