+++ deployment/environment/memory/sui_chains.go
@@ -7,18 +7,12 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/block-vision/sui-go-sdk/models"
-	"github.com/go-resty/resty/v2"
 	"github.com/stretchr/testify/require"
 
 	chainsel "github.com/smartcontractkit/chain-selectors"
 
 	cldf_chain "github.com/smartcontractkit/chainlink-deployments-framework/chain"
-	cldf_sui "github.com/smartcontractkit/chainlink-deployments-framework/chain/sui"
 	cldf_sui_provider "github.com/smartcontractkit/chainlink-deployments-framework/chain/sui/provider"
-	"github.com/smartcontractkit/chainlink-testing-framework/framework"
-
-	"github.com/smartcontractkit/chainlink/v2/core/services/chainlink"
 )
 
 func getTestSuiChainSelectors() []uint64 {
@@ -63,37 +57,4 @@ func GenerateChainsSui(t *testing.T, numChains int) []cldf_chain.BlockChain {
 
 	t.Logf("Created %d Sui chains: %+v", len(chains), chains)
 	return chains
-}
-
-func createSuiChainConfig(chainID string, chain cldf_sui.Chain) chainlink.RawConfig {
-	chainConfig := chainlink.RawConfig{}
-
-	chainConfig["Enabled"] = true
-	chainConfig["ChainID"] = chainID
-	chainConfig["NetworkName"] = "sui-localnet"
-	chainConfig["NetworkNameFull"] = "sui-localnet"
-	chainConfig["Nodes"] = []any{
-		map[string]any{
-			"Name": "primary",
-			"URL":  chain.URL,
-		},
-	}
-
-	return chainConfig
-}
-
-func FundSuiAccount(url string, address string) error {
-	r := resty.New().SetBaseURL(url)
-	b := &models.FaucetRequest{
-		FixedAmountRequest: &models.FaucetFixedAmountRequest{
-			Recipient: address,
-		},
-	}
-	resp, err := r.R().SetBody(b).SetHeader("Content-Type", "application/json").Post("/gas")
-	if err != nil {
-		return err
-	}
-	framework.L.Info().Any("Resp", resp).Msg("Address is funded!")
-
-	return nil
 }
