--- deployment/internal/nodetestutils/node.go
@@ -1,1 +1,0 @@
-package nodetestutils
