+++ deployment/keystone/changeset/test/dons.go
@@ -5,7 +5,7 @@ import (
 	"sort"
 
 	"github.com/smartcontractkit/chainlink/deployment"
-	"github.com/smartcontractkit/chainlink/deployment/environment/memory"
+	"github.com/smartcontractkit/chainlink/deployment/utils/nodetestutils"
 	"github.com/smartcontractkit/chainlink/v2/core/services/keystore/keys/p2pkey"
 )
 
@@ -22,10 +22,10 @@ var _ testDon = (*memoryDon)(nil)
 // memoryDon is backed by in-memory nodes running the full chainlink Application
 type memoryDon struct {
 	name string
-	m    map[string]memory.Node
+	m    map[string]nodetestutils.Node
 }
 
-func newMemoryDon(name string, m map[string]memory.Node) *memoryDon {
+func newMemoryDon(name string, m map[string]nodetestutils.Node) *memoryDon {
 	return &memoryDon{name: name, m: m}
 }
 
@@ -133,8 +133,8 @@ func (d *memoryDons) P2PIDs() P2PIDs {
 	return out.Unique()
 }
 
-func (d *memoryDons) AllNodes() map[string]memory.Node {
-	out := make(map[string]memory.Node)
+func (d *memoryDons) AllNodes() map[string]nodetestutils.Node {
+	out := make(map[string]nodetestutils.Node)
 	for _, d := range d.dons {
 		maps.Copy(out, d.m)
 	}
